pyproject.toml
[tool.poetry]
name = "gpt-va"
version = "0.1.0"
description = ""
authors = ["AlbertoV5 <58243333+AlbertoV5@users.noreply.github.com>"]
readme = "README.md"
# packages = [{include = "gpt_va"}]

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
torchaudio = "^2.0.1"
pyaudio = "^0.2.13"
boto3 = "^1.26.98"
pydantic = "^1.10.7"
pyyaml = "^6.0"
toml = "^0.10.2"
rpp = {git = "https://github.com/Perlence/rpp.git"}
pgvector = "^0.1.6"
sqlalchemy = "^2.0.9"
asyncpg = "^0.27.0"
rich = "^13.3.5"
openai = "^0.27.7"
pyperclip = "^1.8.2"
transformers = "^4.29.2"
mistletoe = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.3.1"
anyio = "^3.7.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
psycopg2 = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] (%(filename)s:%(lineno)s): %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "tests/tests.log"
log_file_level = "INFO"
log_file_format = "%(asctime)s [%(levelname)s] (%(filename)s:%(lineno)s): %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.black]
line-length = 99
target-version = ['py310']